requirements.txt
pdf2docx==0.5.8
pdfplumber==0.11.0
PyPDF2==2.12.1
python-pptx==0.6.18
